requirements.txt
requests~=2.32.5
PyMySQL~=1.1.2
allure-python-commons~=2.15.3
pytest~=9.0.2